requirements.txt
pylint
pillow
moviepy
flask
waitress
python-dotenv

requests==2.32.3
PySocks==1.7.1
pydantic==2.8.2
moviepy==1.0.3
pycryptodomex==3.20.0
